+++ tests/modules/ngx-typeahead.component.spec.ts
@@ -1,17 +1,17 @@
-import { TestBed, ComponentFixture, inject } from "@angular/core/testing";
-import { HttpClientModule, HttpClient } from "@angular/common/http";
+import { TestBed, ComponentFixture, inject } from '@angular/core/testing';
+import { HttpClientModule, HttpClient } from '@angular/common/http';
 import {
   ChangeDetectorRef,
   ElementRef,
   ViewContainerRef,
   DebugElement
-} from "@angular/core";
-import { NgxTypeAheadComponent } from "../../src/modules/ngx-typeahead.component";
-import { NgxTypeaheadModule } from "../../src/modules/ngx-typeahead.module";
-import { Key } from "../../src/models";
-import { Subject } from "rxjs/Subject";
+} from '@angular/core';
+import { NgxTypeAheadComponent } from '../../src/modules/ngx-typeahead.component';
+import { NgxTypeaheadModule } from '../../src/modules/ngx-typeahead.module';
+import { Key } from '../../src/models';
+import { Subject } from 'rxjs';
 
-describe("A Typeahead component", () => {
+describe('A Typeahead component', () => {
   let component: NgxTypeAheadComponent;
   let fixture: ComponentFixture<NgxTypeAheadComponent>;
   // let de: DebugElement;
@@ -23,21 +23,21 @@ describe("A Typeahead component", () => {
   // register all needed dependencies
   beforeEach(() => {
     // setting mocked providers
-    spyElementRef = jasmine.createSpyObj("spyElementRef", ["nativeElement"]);
-    spyViewContainerRef = jasmine.createSpyObj("spyViewContainerRef", [
-      "createEmbeddedView"
+    spyElementRef = jasmine.createSpyObj('spyElementRef', ['nativeElement']);
+    spyViewContainerRef = jasmine.createSpyObj('spyViewContainerRef', [
+      'createEmbeddedView'
     ]);
-    spyHttp = jasmine.createSpyObj("spyHttp", ["get", "post"]);
-    spyChangeDetectorRef = jasmine.createSpyObj("spyChangeDetectorRef", [
-      "markForCheck"
+    spyHttp = jasmine.createSpyObj('spyHttp', ['get', 'post']);
+    spyChangeDetectorRef = jasmine.createSpyObj('spyChangeDetectorRef', [
+      'markForCheck'
     ]);
 
     // setting spy on methods
-    ["filterEnterEvent", "listenAndSuggest", "navigateWithArrows"].forEach(
+    ['filterEnterEvent', 'listenAndSuggest', 'navigateWithArrows'].forEach(
       method =>
         spyOn(NgxTypeAheadComponent.prototype, method as any).and.callThrough()
     );
-    spyOn(NgxTypeAheadComponent.prototype, "renderTemplate");
+    spyOn(NgxTypeAheadComponent.prototype, 'renderTemplate');
 
     TestBed.configureTestingModule({
       declarations: [NgxTypeAheadComponent],
@@ -55,48 +55,48 @@ describe("A Typeahead component", () => {
     fixture.detectChanges();
   });
 
-  it("should have an instance", () => {
+  it('should have an instance', () => {
     expect(component).toBeDefined();
   });
 
-  describe("Default Inputs", () => {
-    it("should have a default taQueryParam", () => {
-      expect(component.taQueryParam).toBe("q");
+  describe('Default Inputs', () => {
+    it('should have a default taQueryParam', () => {
+      expect(component.taQueryParam).toBe('q');
     });
 
-    it("should have a default undefined taCallbackParamValue", () => {
+    it('should have a default undefined taCallbackParamValue', () => {
       expect(component.taCallbackParamValue).toBeUndefined();
     });
 
-    it("should have a default taApi", () => {
-      expect(component.taApi).toBe("jsonp");
+    it('should have a default taApi', () => {
+      expect(component.taApi).toBe('jsonp');
     });
 
-    it("should have a default taApiMethod", () => {
-      expect(component.taApiMethod).toBe("get");
+    it('should have a default taApiMethod', () => {
+      expect(component.taApiMethod).toBe('get');
     });
   });
 
-  describe("Component Init", () => {
-    it("should create a filter enter event subscription", () => {
+  describe('Component Init', () => {
+    it('should create a filter enter event subscription', () => {
       expect(component.filterEnterEvent).toHaveBeenCalledTimes(1);
     });
 
-    it("should create a listen and suggest subscription", () => {
+    it('should create a listen and suggest subscription', () => {
       expect(component.listenAndSuggest).toHaveBeenCalledTimes(1);
     });
 
-    it("should create a navigation with arrows subscription", () => {
+    it('should create a navigation with arrows subscription', () => {
       expect(component.navigateWithArrows).toHaveBeenCalledTimes(1);
     });
 
-    it("should render template", () => {
+    it('should render template', () => {
       expect(component.renderTemplate).toHaveBeenCalled();
     });
   });
 
-  describe("Functionality", () => {
-    it("should hide suggestion when ESC is pressed", () => {
+  describe('Functionality', () => {
+    it('should hide suggestion when ESC is pressed', () => {
       const mockedEvent = {
         keyCode: Key.Escape,
         preventDefault: () => undefined
@@ -108,29 +108,29 @@ describe("A Typeahead component", () => {
     });
   });
 
-  describe("Requests", () => {
+  describe('Requests', () => {
     [
       {
-        title: "should request with http get",
-        method: "get",
+        title: 'should request with http get',
+        method: 'get',
         inputs: {
-          taUrl: "mocked",
-          taApi: "http"
+          taUrl: 'mocked',
+          taApi: 'http'
         },
-        expected: "get"
+        expected: 'get'
       },
       {
-        title: "should request with http post",
-        method: "post",
+        title: 'should request with http post',
+        method: 'post',
         inputs: {
-          taUrl: "mocked",
-          taApi: "http"
+          taUrl: 'mocked',
+          taApi: 'http'
         },
-        expected: "post"
+        expected: 'post'
       }
     ].forEach(({ title, method, inputs, expected }) => {
       it(title, () => {
-        const query = "something";
+        const query = 'something';
         component.taUrl = inputs.taUrl;
         component.taApi = inputs.taApi;
         component.taApiMethod = method;
@@ -139,27 +139,27 @@ describe("A Typeahead component", () => {
       });
     });
 
-    it("should show a static list when given", () => {
-      component.taList = ["aa", "ab", "ac"];
-      const query = "a";
+    it('should show a static list when given', () => {
+      component.taList = ['aa', 'ab', 'ac'];
+      const query = 'a';
       const actual = component.suggest(query);
       actual.subscribe(list => {
-        expect(list.join("")).toMatch(component.taList.join(""));
+        expect(list.join('')).toMatch(component.taList.join(''));
       });
     });
 
-    it("should show a filtered static list when given", () => {
-      component.taList = ["aa", "bb", "cc"];
-      const query = "a";
+    it('should show a filtered static list when given', () => {
+      component.taList = ['aa', 'bb', 'cc'];
+      const query = 'a';
       const actual = component.suggest(query);
       actual.subscribe(list => {
-        expect(list.join("")).toMatch(component.taList[0]);
+        expect(list.join('')).toMatch(component.taList[0]);
       });
     });
 
-    it("should show an empty static list when query not exist in results", () => {
-      component.taList = ["aa", "bb", "cc"];
-      const query = "x";
+    it('should show an empty static list when query not exist in results', () => {
+      component.taList = ['aa', 'bb', 'cc'];
+      const query = 'x';
       const actual = component.suggest(query);
       actual.subscribe(list => {
         expect(list.length).toBe(0);
@@ -167,9 +167,9 @@ describe("A Typeahead component", () => {
     });
   });
 
-  describe("Subscriptions", () => {
-    it("should unsubscribe from all subscriptions", () => {
-      spyOn(component, "handleSelectSuggestion");
+  describe('Subscriptions', () => {
+    it('should unsubscribe from all subscriptions', () => {
+      spyOn(component, 'handleSelectSuggestion');
       component.ngOnDestroy();
       const event = {
         keyCode: Key.Enter
@@ -180,12 +180,12 @@ describe("A Typeahead component", () => {
 
     [
       {
-        title: "should set suggestion index to start",
+        title: 'should set suggestion index to start',
         index: 0,
         expected: true
       },
       {
-        title: "should NOT set suggestion index to start",
+        title: 'should NOT set suggestion index to start',
         index: 3,
         expected: false
       }
@@ -193,16 +193,16 @@ describe("A Typeahead component", () => {
       it(title, () => {
         const demo$ = new Subject<any>();
         component.taList = [
-          "guitar",
-          "drums",
-          "bass",
-          "electric guitars",
-          "keyboards",
-          "mic"
+          'guitar',
+          'drums',
+          'bass',
+          'electric guitars',
+          'keyboards',
+          'mic'
         ];
         component.listenAndSuggest(demo$);
-        demo$.next({ target: { value: "testing" } });
-        const actual = component.markIsActive(index, "testing");
+        demo$.next({ target: { value: 'testing' } });
+        const actual = component.markIsActive(index, 'testing');
         expect(actual).toBe(expected);
       });
     });